provider: use fallback provider helper
Allow the DTL to sync from multiple web3 backends using
the `ethers.providers.FallbackProvider`. Multiple URLs
can be passed through as a comma delimited string to
configure multiple providers.

A helper function is added to `core-utils` to allow
easy configuration between services. This PR only introduces
the usage of the fallback provider into the DTL. In the
future, it should be added to other services that query L1
as well.