l2geth: no longer sync batches as sequencer
Previously the sequencer would call the `syncBatchesToTip` method
as part of its `sequence()` loop. This commit removes that as the
call to `syncBatchesToTip` does nothing for the sequencer currently.
Syncing the batches would be useful for triggering reorgs, but
that functionality does not currently exist. It would be better to
remove this functionality for now as it could potentially interfere
with the locking mechanism introduced in
dfaa8fc20f586265f0a5c07ca76c6c793ba49157.

This will also reduce the amount of time the `txLock` is held, meaning
that less transactions sent via RPC will be blocked and held in memory
while waiting to execute.