l2geth: fix various tests
The tests were not running in CI on the experimental
branch which was an oversight and allowed for various
tests to fail. Now that experimental has been renamed
to regenesis/0.5.0, the tests are running in CI again.
This commit fixes all tests that were broken while
development happened on experimental.

It removes a config option `--rollup.minl2gaslimit`
that is no longer required post regenesis.

It also adds the ability to populate a genesis account
with value in the sync service tests.