batch-submitter: trigger release
Trigger a release of the batch submitter with the yatm fix.
The batch submitter would get stuck in a loop on kovan,
attempting to submit a transaction with a bad nonce over
and over again. The fix was to add a hook function
that can notice when there is a nonce error.