fix: check for empty block before submitting task in miner
Prior to this commit, an empty block would be submitted to the taskCh
before sanity check whether or not it was empty. This assertion is moved
up to fail before submitting the task.

It seems the placement of the check before was done to get around the
StreamUncle unit test, which would allow commit to submit an empty task
before ultimately failing. Now that the test is modified to reflect the
expected behavior of the sequencer, this placement is no longer
necessary.