fix: disable geth preemptive empty block mining logic
When using PoW, geth begins mining on empty blocks before attempting to
select which transactions to include. This optimization is not needed,
as we require that every L2 block have exactly one transaction. This
commit disables this behavior an modifies the tests accordingly. The
test is also expanded to be stricter wrt to testing number of txs and
uncles on all blocks.