chore: add changeset for l2geth
Add missing changeset for https://github.com/ethereum-optimism/optimism/pull/2011

A replica can be configured with the flag
`--sequencer.clienthttp` or the env var
`SEQUENCER_CLIENT_HTTP` which points to the
sequencer and when users hit the RPC endpoint
`eth_sendRawTransaction` it will forward the
value to the sequencer.

If running a mainnet node, this value should be
set to `https://mainnet.optimism.io` or an
infrastructure provider such as Quiknode, Infura
or Alchemy.