fix: update batch submitter to include all ts info
This commit fixes a bug in the batch submitter where not all timestamp
information was being correctly rolled up. This has temporarily caused
certain L1 syncing nodes to generate an incorrect state root. This PR
will be followed by a second PR with an update to the DTL to guarantee
that incorrect timestampd data can be corrected.