dtl: enable typed batch support
The data transport layer will now be able to index
typed batches. A typed batch has the first batch context
have a timestamp of 0 and the blocknumber is used as an
enum to determine the type of the batch.

It is not possible to have a timestamp of 0 in realistic
conditions. Batches that have a non zero timestamp
as the first batch context are considered a legacy batch.

The first typed batch is type 0 where the blocknumber is 0
and the transaction data is compressed with zlib.