feat: always run go unit tests on every PR
The current methodology only attempts to run unit tests when changes to
the package paths are detected. However, this is brittle and most
notably doesn't capture when changes are made to the workflow itself. As
a result, a green build in a PR doesn't neccessarily correlate to a
green build after merge. Unit tests are cheap, so this should prevent
needless false positives.