proxyd: Don't hit Redis when the out of service interval is zero
When the out of service interval is zero, `proxyd` will send a `SETEX` command to Redis with a zero expiry. This is technically an error, and causes messages like `error setting backend unavailable ERR invalid expire time in setex` to appear in the logs. Users won't notice the issue since `proxyd` can continue working when Redis returns an error, but we should clean this up nonetheless since these problems appear in our alerting.