fix(rhc): fix bug in healthcheck dockerfile
Fixes a bug in the replica-healthcheck dockerfile. Also includes
replica-healthcheck as part of the ops setup by default to make this
stuff easier to catch and debug in the future. We will likely still want
to add integration tests that query the metrics defined in the
replica-healthcheck to confirm that everything is working as expected.