ci: Use Alpine to reduce image size
Image sizes were increased because the packages Dockerfile used `buster-slim` rather than Alpine. Unfortunately, the images will still be large because the `node_modules` directory contains dev dependencies. This is how it was before the CircleCI migration/larger build cleanup, and will need a larger rethink of how we build TypeScript projects in the monorepo before it can be solved.