feat(ci): check deploy markdown correctness (#2373)
* feat(ci): check deploy markdown correctness

Adds a step to the contract tests to confirm that the contract
deployment docs are up to date. Will only fail if a deployment has been
modified but yarn autogen:markdown was not run properly.

* maint(ct): update deployments markdown

Updates the deployments markdown files. Should've been updated after we
deployed the Teleportr contracts to Kovan but we didn't have the CI
workflow in place to detect that this didn't happen.