op-e2e: test l2 fee redirect; l1 availability fee (#2770)
This PR depends on #26 on reference-optimistic-geth:

https://github.com/ethereum-optimism/reference-optimistic-geth/pull/26

A new test TestFees is added, which runs a simple transaction and tests
that the L2 fee is redirected to the configured fee recipient and that
the L1 availability fee is calculated based the RLP encoded transaction
size and also redirected to the configured fee recipient.

TestFees is skipped, but can be enabled the above PR is merged into
reference-optimistic-geth.

NOTE: After #26 reference-optimistic-geth is merged, L2 genesis
configuration needs to be updated for OptimismConfig to enable special
fee handling and to find L1Block, GasPriceOracle contract addresses.