contracts-bedrock: better logging in deposit task
This task is ran in CI and it started failing recently
for unknown reasons. This increases the logging so
that it is easier to understand what is going on
in the script. It detects the case where the deposit
goes through but is undetected by client side hash
computation. It also periodically logs the block number
and hash of the latest block so it can be known if there
has been a reorg.