style(ctp): move TeleportrWithdrawer to L1 folder (#2995)
Minor change, moves the TeleportrWithdrawer contract to the L1 folder to
keep in line with the placing of other Teleportr contracts. We currently
only use this contract on the L1 side. If we ever want an L2 to L1
teleportr, we might want to put all of these things in universal.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>