op-node: Properly set timestamp in PreparePayloadAttributes (#3151)
This now uses a passed in timestamp for PreparePayloadAttributes.
This is important because when generating a payload attributes from
a batch it should be using the timestamp of the batch, not the
L2 parent + block time. When we are creating payloads based on the
L2 parent time, that is easy to set there instead of pulling it into
this function.

This enabled/exacerbated a divergence bug where the timestamp of the
batch was not the timestamp that was being set inside
PreparePayloadAttributes.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>