op-node: Fixed Integers in channel frame header (#3126)
* goals for fixed int

* op-node: Fixed integer sizes in the channel header

This mainly modifies the channel_frame, but has some knock on effects
as the width of some of the fields have been reduced. The channel
frame code is also changed more than I expected due to differences
in API of working with fixed int vs uvarints in go. Otherwise the
code reads very similarly with using Reader/Writer APIs.

* op-node: fix frame unmarshal func to return correct err

* specs: update frame format specs

* Update op-node/rollup/derive/channel_bank_test.go

Co-authored-by: protolambda <proto@protolambda.com>
Co-authored-by: Matthew Slipper <me@matthewslipper.com>