deps: remove extra oz deps (#3504)
There were open zeppelin contracts defined in the top level
`package.json` as well as in each of the individual packages.
We want to be very explicit with the version of the contract
dependencies that are used. I believe that the versions local
to the packages were being used instead of the globally defined
versions, removing them from the top level `package.json` will
ensure this.

Co-authored-by: Matthew Slipper <me@matthewslipper.com>