op-node: Fix reset bug in batch queue (#3694)
The batch queue was using the previous origin when computing if
the origin was behind or not. It would then immediately advance
it's internal origin to the next origin (confusingly coming from
the `prev` stage). It would take action on the current origin based
on out of date data. Specifically, it would not save an origin that
should have been saved into the l1Blocks array. This causes errors
because the L1 Origin of the safe head would not be present when
attempting to derive the next batch after a reset.