op-node: Skip conf depth check if L1 head is empty
The L1 Head block in the conf depth is pulled from the L1 State object.
The L1 State is never initialized at startup but instead from the new head
channel. This means that prior to the first new head, any code which calls
BlockByNumber would always fail if the conf depth was not zero.