feat(dtl): handle deposit shutoff
Updates the DTL to correctly handle the deposit shutoff flag. We put the
flag inside the AddressManager because the DTL already had easy access
to the AddressManager. We parse the address as a uint256 and check if
it's non-zero. By checking this every loop we can smoothly recover if
the shutoff flag is reset.