op-chain-ops: various cleanup
Fixes a bug were an error should have been returned and it was not.
Wraps some errors so that its easier to tell what the problem is
in the case of there being an error. A bit of refactoring that
allows the strict withdrawal migration checks to be skipped optionally.
This should not be used in production, but can be used in early test
iterations.