op-chain-ops: more cleanup in migrate script
The `op-migrate` script had some leftover stuff
in it from early iterations on its development.
The leftover stuff was cleaned up to improve the
overall UX of the script. Now code should not need
to be commented out + rebuild the binary, it will
automatically detect when the script has already
ran on a datadir.

Also pass through the no check arg so that we can
re-enable the checking and dynamically change if
we want to run the data validation as part of the
migration.