feat(ctb): break out system dictator steps (#4553)
Breaks out the SystemDictator steps into two files so that the steps
that get executed before the system is shut down can easily be separated
from the steps that executed after the system gets spun back up. Idea
here is that you would execute with the tag "steps1", wait for the
legacy system to come down, migrate the L2, then execute with the tags
"steps2" to complete the upgrade.