op-chain-ops: do not allow 0 scalar
The `DeployConfig` value for the `GasPriceOracleScalar` should
not be allowed to be set to 0. If it is set to 0, it would effectively
make the L1 portion of the fee nothing. This isn't useful since
eip1559 is enabled which required ether to be burnt with
execution, meaning that free transactions are not allowed
on the network.

If this value is set to 0, it is most likely a bug in the config
and we do not want to lose money by charging no L1 portion of the
fee.