op-chain-ops: always use magic for transition/genesis block
No longer allow for a configurable L2 genesis/transition
block extradata. The extradata will always be set to
`BEDROCK`. This allows us to have automation to ensure that
the value set as the `L2OutputOracle.startingBlockNumber`
matches the actual L2 genesis/transition block.

Note that the term genesis block and transition block are
used interchangably, the term transition block is used
for upgraded networks.