contracts-bedrock: remove storage spacers gpo
Removes the storage spacers for the `GasPriceOracle`
since the migration will be wiping the storage layout.

Must be merged with https://github.com/ethereum-optimism/optimism/pull/4605

The fee values were stored in the `GasPriceOracle` predeploy
before the bedrock upgrade. Afterwards they are in the `L1Block`
contract. This simplifies the `GasPriceOracle`.