ctp: delete hardhat drippie tests
The foundry tests merged in https://github.com/ethereum-optimism/optimism/pull/4760
are a superset of the hardhat tests so we can delete the
hardhat drippie tests in favor of the foundry based tests.
This is moving us closer to a unified smart contract
development environment where we can enforce the same standards
across all smart contracts developed in the monorepo.