txmgr: Don't enforce price bump if new price is not larger
This handles the case that a transaction is being resubmitted while
the L1 Basefee & Tip have not increased. In this case we reuse the
previous values. If the new values are between the old values & the
threshold, we then increase that value to the threshold. If it is
larger than the threshold, we then use the suggested value.

We need to not always bump the transaction price to ensure that we
don't go exponential when resubmitting transactions.