contracts-bedrock: L2StandardBridge legacy `withdraw`
Adds a test for the legacy withdraw functionality on the
L2StandardBridge. Previously there were no tests for
withdrawing ether via `withdraw`. This adds coverage.
Also natspec is added to describe that these functions
are not useful for being able to withdraw native L2 tokens.
The modern functions must be used to withdraw native
L2 tokens, the legacy functions assume that the withdrawal
is working on a OptimismMintableERC20 token.

Also regenerate the bindings for the slight change to the
bytecode where we save a call in the withdrawal path
if we don't need to make the call.