contracts-bedrock: prevent deposits with too small gaslimit
The deposits sent to L2 should have a gas limit that at least
covers the intrinsic gas. If the deposits do not pay that much,
then it is possible for cause resource usage on L2 without
paying for it. This is not a critical bug because the user
would still need to pay for L1 inclusion.