batch-submitter: allow deposit only batches
Fixes a long standing bug that allows the batch submitter
to send batches that only contain deposit transactions.
Includes test coverage of the batch creation.

The on chain code requires that there must be at least one
context. A context contains information about the queue
transactions that should be included. It makes no sense
to error when there is a context but no sequencer txs
because the context must be submitted to the chain.