op-chain-ops: prevent invalid decoding errors
Invalid decoding errors should not stop the script from
progressing. This fix ensures that all possible errors
are handled in a way that they can be skipped. The panic
conditions should never happen, so we do not need to worry
about them. It is permissionless to insert data into the
set that is parsed through this function, so errors cannot
break it.

Also adds unit tests for the parsing, ensures to cover
each relevant branch.