contracts-bedrock: update portal min gas limit
Prevent users from sending cross domain messages with too small
of a gas limit. Users must pay for resources they consume. The
L2 network must store deposit transactions in their history and
they are also included in the p2p gossip of unsafe blocks.
A linearly increasing requirement based on the calldata size
is introduced. A base of 21k minimum gas limit is required when
there is no gas limit, the formula `calldata.length * 16 + 21000`
is used to enforce a min gas limit.

The larger the gaslimit, the more ether that is burnt on L1.