fix(ci): use npm@8.11.0
Updates the ci builder Dockerfile to use a pinned version of npm that is
known to not have a module resolution bug found in the default version
of npm that ships with the node version from setup_16.x. Resolves an
issue when importing packages that have hardhat as a dependency, which
causes the current version of npx to try to execute the hardhat binary
found deep within the dependency tree, which in turn causes hardhat to
throw a fit.